Renames cli package into core
