+++ pkg/core/cli.go
@@ -1,4 +1,4 @@
-package cli
+package core
 
 import (
 	"fmt"
