+++ pkg/core/cli_test.go
@@ -1,4 +1,4 @@
-package cli
+package core
 
 import (
 	"context"
