+++ pkg/core/context.go
@@ -1,4 +1,4 @@
-package cli
+package core
 
 import (
 	"io"
