+++ pkg/core/input.go
@@ -1,4 +1,4 @@
-package cli
+package core
 
 import "github.com/eiannone/keyboard"
